Use request location if locname has unsupported characters
Sometimes, wttr.in will use a more specific or accurate weather location name than the request. This location name may contain unsupported characters (like characters from non-English languages). This will cause an error while encoding. To prevent this, the weather request function will check if there's any error with encoding it. If there is an error, it will use the request location, not the wttr.in provided weather location.